beta 2.0
This was the best successful attempt when dealing with multiple songs, after 3 days i got a way to change songs without it playing together with others (a complete mess) when clicking on a div.

Before clicking it paused the song, now it plays from the beginning or plays another one if you choose. What is joyful on the one hand, is completely frustrating on the other, this brought me a problem that I didn't have, which is not being able to pause the songs when clicking the play button.